Rework snippet and sl4a startup process.
_is_app_running() didn't work before because the socket was 'live' as soon
as it was bound through adb, even though nothing had started on the other
end. This led to race conditions between the snippet app starting and the
client connecting.

I couldn't find a way to make sure something was alive on the other end
other than actually sending data through the socket and getting a response,
so I've changed the is_app_running() process to actually call connect().
Tested with both snippet and sl4a.

Fixes #88.