Logcat publisher/subscriber service.
Implement feature described in https://github.com/google/mobly/issues/486.
This implementation makes use of the AndroidDevice Service Manager and
lazy instantiates a logcat publisher as a `logcat_publisher` singleton member
upon registering a subscriber.

Users are free to define their own subclass of LogcatSubscriber and
register that subscriber with the publisher. Mobly includes one example
of a LogcatSubscriber called LogcatEventSubscriber and is used as a
context manager to track events within logcat. For example:

  >>> from mobly.controllers import android_device
  >>> ad = android_device.AndroidDevice('0123456789')
  >>> with ad.logcat_event(pattern='writeEvent (.*) (.*)') as event:
         event.wait()
         print('time: %s' % event.trigger.time)
         print('pid: %d' % event.trigger.pid)
         print('tid: %d' % event.trigger.tid)
         print('tag: %s' % event.trigger.tag)
         print('message: %s' % event.trigger.message)
         print('regex match groups: %s' % str(event.match.groups()))
  time: 2018-10-02 16:19:00.408000
  pid: 1463
  tid: 1463
  tag: vol.Events
  message: writeEvent dismiss_dialog volume_controller
  regex match groups: ('dismiss_dialog', 'volume_controller')