record: emit explicit end when appending to summary
Since the `yaml` summary file is appended during the process, a reader
is able to handle any new record as soon as they have been written. But
it may read incomplete `yaml` data, for example on Linux there is no
guaranty that a single write to file will be available as one read.

By explicitly emitting the end marker to the record, the reader may now
be able to read the complete `yaml` data before processing it.