record: add `TestClassList` to emitted records
When running multiple test classes in the test runner, it actually emit
multiple `TestNameList` but one `Summary`. Which make impossible for a
reader of the summary file to retrieve the complete list of tests before
they had been executed.

To solve this a `TestClassList` record is now emitted before the run,
which contains the full list of tests in the following format:
```yaml
Classes:
- Name: Foo
  Tests:
  - test_a
  - test_b
- Name: Bar
  Tests:
  - test_c
  - test_d
```

It was chosen to add a new record and to to not change the actual
`TestNameList` record to not break the actual behavior.