Add `get_service_alias_by_class` in Android service manager.
This is useful for looking up how many instances of a service has
been registered, and find out their aliases.

Util functions may depend on services like `uiautomator`. Users tend
to follow the anti-pattern of attempting to register the service
again under the name the util expects. This causes duplication.
This API can be used to clean things up.