wait-for-disconnect and then wait-for-device
Previously, right after we do `adb reboot`, we immediately executed `adb
wait-for-device`. This is problematic because the device may still be in
the process of shutting down (shutdown itself is taking some time), in
which case `adb wait-for-device` would immediately return, giving a
false signal that the reboot was finished.

Fixing this issue by firstly waiting for the adb connection to be lost
(via adb wait-for-disconnect) and when executing adb-wait-for-device.

Note that other test hardness like tradefed already has had this. ex:
https://android.googlesource.com/platform/tools/tradefederation/+/refs/heads/main/src/com/android/tradefed/device/NativeDevice.java#4023

Test: tox passed